fixed plotter:

+++ setup.py
@@ -97,7 +97,7 @@ class custom_install(install):
 
     # Default compiler and precision level for the main openmoc module
     self.cc = 'gcc'
-    self.fp = 'single'
+    self.fp = 'double'
     self.mpi = True
 
     # Set defaults for each of the newly defined compile time options
